requirements.txt
micropyGPS==1.0
pyserial==3.4
pywin32==224
